:sparkles: feat: create userService test
